+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - the texture pointer of a tilemap is now sent only using `.use` method
 - disabling warning when converting float to int, int to float in the tilemap/tileview
 - all the sizes in the tileview are now using `sf::Vector2i`
+- all the sizes in the tilemap are now using `sf::Vector2i`
 
 ### Removed
 - removed the argument `sgl::SceneManager&` from the scene, now using an internal pointer to the scene manager, in the scene itself